% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins_bundle_code.R
\name{bundle_addin}
\alias{bundle_addin}
\alias{bundle_code}
\title{Bundle selected code add-in}
\usage{
bundle_code()
}
\description{
This add-in will bundle the selected R code into a file you choose.

A file dialogue box prompting you to select the output file. When it is saved, the file will pop up.
}
