% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_backend.R
\name{clipboard_is_numeric}
\alias{clipboard_is_numeric}
\title{Is the clipboard numeric?}
\usage{
clipboard_is_numeric(clipboard = get_clipboard())
}
\arguments{
\item{clipboard}{A character vector (defaults to the clipboard)}
}
\value{
A \link{logical} vector (\code{TRUE} or \code{FALSE}) of length 1.
}
\description{
A utility function that performs a logical operation to check if the clipboard contains only numbers.
}
