% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_backend.R
\name{clipboard_utilities}
\alias{clipboard_utilities}
\alias{get_clipboard}
\alias{set_clipboard}
\title{Read and write to the system clipboard}
\usage{
get_clipboard()

set_clipboard(text)
}
\arguments{
\item{text}{A character vector, each element of which is a line of text to be written to the clipboard.}
}
\value{
The value of \code{set_clipboard()} is a character vector, each element of which is a line of text in the system clipboard.

The value of \code{set_clipboard()} is always \code{\link{NULL}} as it is used for its ``side-effect'' if writing to the clipboard and not for its value.
}
\description{
Utility functions to read and write to the system clipboard with checks.

     These functions are used in the source code for add-ins in the \code{fwdSlash} package.
}
\details{
These functions are essentially the \code{clipr} functions, \code{\link[clipr]{read_clip}} and \code{\link[clipr]{write_clip}} respectively with some additional checks

These checks include:
\itemize{
 \item If the \code{clipr} package is incompatible with the system, it will provide an informative dialogue box.
 \item If the R session is not interactive, these functions will not execute the clipr functions as they are not meant to be used in non-interactive sessions.
}
}
