% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_console_output.R
\name{console_output_method}
\alias{console_output_method}
\title{Format console output}
\usage{
console_output_method(obj)
}
\arguments{
\item{obj}{An R object to be changed formatted according to its class. It will default to \link[utils:capture.output]{the default caupture output} for that class.}
}
\description{
Utility function for changing objects to a format that can then be pasted using special copy and paste add-ins for console output.
}
