% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_file_paths.R
\name{convert_file_path}
\alias{convert_file_path}
\alias{R_to_sys_path}
\alias{sys_path_to_R}
\title{Convert between file paths}
\usage{
R_to_sys_path(path)

sys_path_to_R(path)
}
\arguments{
\item{path}{A file path in a form appropriate for R in \code{R_to_sys_path()} and in a form appropriate for the operating system in \code{sys_path_to_R()}.}
}
\description{
Use these functions to convert between R file paths and system file paths. These function make the most sense on a Windows machine since R uses the same file paths as Unix-based systems (i.e. Mac and Linux).

It will also try to convert relative paths to absolute paths if possible.
}
