% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_addins.R
\name{copy_add-ins}
\alias{copy_add-ins}
\alias{copy_f_to_b}
\alias{copy_f_to_bb}
\alias{copy_bb_to_b}
\title{Special copy add-ins}
\usage{
copy_f_to_b()

copy_f_to_bb()

copy_bb_to_b(x)
}
\description{
Use these add-ins to copy the current selected text to the cliboard with some adjustments to slashes (`\`, `\\` and `/`).
}
\details{
See the table of correspondence between function names and add-in names below.
   \tabular{ll}{
   \strong{Function} \tab \strong{Add-in} \cr
   \code{copy_f_to_b()} \tab Special copy (/ to \\) \cr
   \code{copy_f_to_bb()} \tab Special copy (/ to \\\\) \cr
   \code{copy_bb_to_b()} \tab Special copy (\\\\ to \\) \cr
   }
}
