% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins_file_paths.R
\name{file_path_add-ins}
\alias{file_path_add-ins}
\alias{copy_file_path}
\alias{paste_file_path}
\title{Special paste slash add-ins}
\usage{
copy_file_path(text = get_selected_text())

paste_file_path(text = get_clipboard())
}
\arguments{
\item{text}{Either the \link[clipr:read_clip]{clipboard}, \link[rstudioapi:rstudio-editors]{current text selection} or a \link{character}.}
}
\description{
Use these add-ins to copy/paste the clipboard to the current cursor location with some adjustments to change file paths.

     Note that they do \strong{not} change the clipboard.
}
\note{
The word ``paste'' here is not to be confused with the base R \code{\link{paste}} function, which is for concatenation.

    The usage here is the more general use meaning of ``paste'', as in placing the current copied item (i.e. in the clipboard) to the cursor position.
}
