% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins_comment_section.R
\name{fill_equals}
\alias{fill_equals}
\title{Fill comment section with equals signs}
\usage{
fill_equals()
}
\description{
This add-in will fill the rest of the line with equals signs in order to break R scripts into sections.
}
