% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_console_output.R
\name{get_call_history}
\alias{get_call_history}
\title{Output the current \code{.Rhistory}}
\usage{
get_call_history()
}
\description{
This is a utility function written specifically for use in the \code{\link[RADins:dot-Last.call]{.Last.call()}} function.
}
