% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins_console_output.R
\name{input_output}
\alias{input_output}
\alias{paste_console_output}
\alias{copy_console_output}
\alias{paste_last_call}
\alias{copy_last_call}
\title{Copy/paste console input/output add-ins}
\usage{
paste_console_output(obj = .Last.value)

copy_console_output(obj = .Last.value)

paste_last_call(...)

copy_last_call(...)
}
\arguments{
\item{obj}{An R object, default to the output of the last command sent to the R console.}

\item{...}{Arguments to pass to \code{\link[fwdSlash:.Last.call]{.Last.call()}}.}
}
\description{
Add-ins that copy/paste the output of the last R evaluation sent to the console. Some attempt is made to adjust the output based on the object. When in doubt, the default \link[utils:capture.output]{capture output} is used.

Add-ins that copy/paste the last R call will get the most recent R code, parse it (i.e. remove comments and potentially remove some of the unnecessary spacing).
}
