% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwdSlash.R
\name{replace_clipboard}
\alias{replace_clipboard}
\alias{replace_from_clipboard}
\alias{replace_to_clipboard}
\title{Replace slashes in the clipboard}
\usage{
replace_from_clipboard(pattern, replace)

replace_to_clipboard(text, pattern, replace)
}
\arguments{
\item{pattern}{A \link[base:regex]{regular expression} to be replaced in the clipboard.}

\item{replace}{A string to replace any instances of argument \code{pattern} with. Ignored if \code{pattern} is an empty string (i.e. \emph{""}).}

\item{text}{Text to be writeen to the clipboard.}
}
\value{
\code{replace_from_clipboard()} returns a character which is a manipulated version of the clipboard.

    \code{replace_to_clipboard()} always returns \code{\link{NULL}} and is used for it's ``side-effect'' (i.e. writing a manipulated version of the argument, \code{text} to the clipboard).

    In both cases, said manipulation is a \link[stringr:str_replace]{regex replacement} specified by the arguments \code{pattern} and \code{replace}.
}
\description{
Utility functions for reading/writing to the clipboard and performing a \link[stringr:tr_replace]{regex replacement} in one step.


    These functions were designed for use in the source code for add-ins that perform \link[fwdSlash:copy_add-ins]{special copies} and \link[fwdSlash:paste_add-ins]{special pastes} with manipulations of slashes in the clipboard in the \code{fwdSlash} package.
}
\note{
Note that this add-in will not manipulate the \code{\link[clipr:read_clip]{clipboard}} itself, only accesses it. This means you can use the clipboard with backlashes in it in another application if need be.
}
