% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwdSlash.R
\name{replace_from_clipboard}
\alias{replace_from_clipboard}
\title{Replace `\` with `/` in the clipboard}
\usage{
replace_from_clipboard(pattern, replace)
}
\arguments{
\item{pattern}{Regular expression to be replaced in the clipboard.}

\item{replace}{What to replace Ignored if \link{pattern} is an empty string (i.e. \code{""}).}
}
\description{
Add-in that will replace all forwardslashes with backslashes. It is intended to be used to copy file paths from the Windows File Explorer.
}
\note{
Note that this add-in will not manipulate the \code{\link[clipr:read_clip]{clipboard}} itself, only accesses it. This means you can use the clipboard with backlashes in it in another application if need be.
}
