% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_backend.R
\name{Rstudio_position_utilities}
\alias{Rstudio_position_utilities}
\alias{get_selected_text}
\alias{write_to_cursor}
\title{Read from/write to the current position in RStudio}
\usage{
get_selected_text()

write_to_cursor(text)
}
\arguments{
\item{text}{A character vector, each element of which is a line of text to be written to the current cursor position.}
}
\value{
The \code{read_selected_text()} function returns a character containing the active selected text in the RStudio IDE.

    \code{write_to_cursor} returns \code{\link{NULL}} as it is the ``side-effects'' (i.e. inserting text in RStudio) that this function is useful for.
}
\description{
Utility functions for reading the current selected text and writing to the current cursor position respectively, by virtue of the \code{\link[rstudioapi:rstudio-documents]{rstudioapi}}.

     These functions are used in the source code for add-ins in the \code{fwdSlash} package.
}
