% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_vector_code.R
\name{to_vector_code}
\alias{to_vector_code}
\title{Convert vector to R code for a vector}
\usage{
to_vector_code(x = clipr::read_clip())
}
\arguments{
\item{x}{A character vector (defaults to the \code{\link[clipr:read_clip]{clipboard}}).}
}
\value{
We show the output by way of the following examples that assume \code{x} is the clipboard.

  \strong{Numeric:}
     suppose we copy the following into the clipboard...

     \code{1}

     \code{2}

     \code{3}

     Then the function will output \code{c(1,2,3)}.

     \strong{Character:}
     suppose we copy the following into the clipboard...

     \code{Apple}

     \code{Orange}

     \code{Banana}

     Then the function will output \code{c("Apple","Orange","Banana")}.
}
\description{
This function converts the clipboard (or indeed any character vector) into the \link[base:c]{code for an R vector}. This can be useful if there it is necessary to ``hard-code'' vectors.

     It treats entirely numeric clipboards --- picked up by \code{\link{clipboard_is_numeric}} --- slightly differently to \emph{supposed} ``character'' clipboards.
}
\examples{
\dontrun{
 # numeric example
 clipboard_to_R_vector(c("1","2","3"))
 # character example
 clipboard_to_R_vector(c("Apple","Orange","Banana")) }

}
