% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins_vector_code.R
\name{View_obj}
\alias{View_obj}
\title{Add-in for viewing objects}
\usage{
View_obj(text = get_selected_text())
}
\arguments{
\item{text}{The currently selected R Code.}
}
\description{
This add-in will \link[base:View]{view} the result of the current selected code. This could be either the name of an object or some code that when run will output an object.
}
