% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_addins.R
\name{copy_add-ins}
\alias{copy_add-ins}
\alias{copy_f_to_b}
\alias{copy_f_to_bb}
\alias{copy_bb_to_b}
\title{Special copy add-ins}
\usage{
copy_f_to_b()

copy_f_to_bb()

copy_bb_to_b(x)
}
\description{
Use these functions to copy the current selected text to the cliboard with some adjustments to slashes (`\`, `\\` and `/`).
}
\details{
See the table of correspondence between function names and clipboard patterns below.
   \tabular{rr}{
   \strong{Function} \tab \strong{Clipboard} \cr
   \code{b} \tab \\ \cr
   \code{bb} \tab \\\\ \cr
   \code{f} \tab / \cr
   }
}
