% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_addins.R
\name{paste_add-ins}
\alias{paste_add-ins}
\alias{paste_b_to_f}
\alias{paste_bb_to_f}
\alias{paste_b_to_bb}
\title{Special paste add-ins}
\usage{
paste_b_to_f()

paste_bb_to_f()

paste_b_to_bb()
}
\description{
Use these add-ins to paste the clipboard to the current cursor location with some adjustments to slashes (`\`, `\\` and  `/`). These add-ins do \strong{not} change the clipboard.
}
\details{
See the table of correspondence between function names and clipboard patterns below.
   \tabular{rr}{
   \strong{Function} \tab \strong{Clipboard} \cr
   \code{b} \tab \\ \cr
   \code{bb} \tab \\\\ \cr
   \code{f} \tab / \cr
   }
}
\note{
The word 'paste' here is not to be confused with the base R \code{\link{paste}} function for concatenation.
}
