% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_change_shortcuts.R
\name{set_keyboard_shortcuts}
\alias{set_keyboard_shortcuts}
\alias{set_shortcuts}
\alias{set_shortcuts_default}
\title{Set keyboard shortcuts}
\usage{
set_shortcuts(addins, shortcuts)

set_shortcuts_default()
}
\arguments{
\item{addins}{Characer vector of underlying functions for addins. Need to be in the right form (e.g. \code{"fwdSlash::copy_file_path"}).}

\item{shortcuts}{Characer vector of keyboard shortcuts for addins. Need to be in the right form (e.g. \code{"Ctrl+Shift+V"}).}
}
\description{
Manipulated the JSON file to make the suggested keyboard shortcuts. Need to close and re-open RStudio (or start a new session) for them to take effect.
}
